filter07/09, esp03/11/15/1b matching; unit_info uses reloc-aware diff percentages; sync_symbols overload resolution via split imports

+++ src/game/dvd.cpp
@@ -680,7 +680,7 @@ void cDvdQueue::readMain()
                 SndMem.blk_mram[t] = Snd.bgm_mram;
                 break;
             }
-            destAddr = SndMem.blk_mram[t];
+            destAddr = (u32) SndMem.blk_mram[t];
             Snd.blk_flag[0] &= ~(1 << t);
             remain = (*ph)->size;
             ofs = hedOfs[depth] + (*ph)->ofs;
